repo police moving utils and generate py files to utils folder
