--- consonance/main.py
@@ -1,2 +1,0 @@
-# code to be added...
-pass
